feature(session): Add session class
Add basic Singleton class for sessions to be used as local storage.
More variables should be added as we develop more features.
As of now the session stores the room link this client will be using (so the moderator link for moderators and the student link for students), the room name and if this user is a moderator or student.